+++ pkg/gadgets/helpers.go
@@ -47,7 +47,7 @@ func CloseLink(l link.Link) link.Link {
 // like node, namespace, pod name and container name when the mount namespace
 // is available.
 type DataEnricherByMntNs interface {
-	EnrichByMntNs(event *types.CommonData, mountnsid uint64)
+	EnrichByMntNs(event *types.CommonData, mountnsid uint32)
 }
 
 // DataNodeEnricher is used to enrich events with Kubernetes node, without
